More renames in the sake of consistency
