+++ ui/DesignSystem/Primitive/Icon/ArrowBoxUpRight.svelte
